add update borrower rating after new opinion added

+++ src/main/java/pl/fintech/solidlending/solidlendigplatform/infrastructure/database/user/OpinionEntity.java
@@ -20,7 +20,7 @@ public class OpinionEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	Long id;
-	Integer opinionRating;
+	Double opinionRating;
 	String opinionText;
 	String author;
 	Long investmentId;
